blog post heading rendering correctly
