+++ app/(posts)/second-post/page.mdx
@@ -6,6 +6,10 @@ export const postDetails = {
   tags: ['tag1', 'tag2']
 }
 
+import PostHeading from '@/app/components/postheading.tsx'
+
+<PostHeading metadata={postDetails} />
+
 ### Second post
 
 This is an example of a second post.
