+++ types.d.ts
@@ -1,10 +1,8 @@
-// export interface PostMetadata {
-//   slug: string;
-//   title: string;
-//   publishDate: string;
-//   tags: string[];
-//   author: string;
-//   publish: boolean;
-//   standfirst: string;
-//   // id: string;
-// }
+export interface PostMetadata {
+  slug: string;
+  title: string;
+  publishDate: string;
+  tags: string[];
+  author: string;
+  standfirst: string;
+}
